pyproject.toml
[tool.poetry]
name = "airbnb-backend"
version = "0.1.0"
description = ""
authors = ["junyoung <jy940608@naver.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "airbnb_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.7"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
